feat(printing): Add printing to imagebase (#1085)
Moving the print function to imageBase will allow
MultiImageViewer to inherit the functionality.
